requirements.txt
streamlit
huggingface-hub
pypdf
python-dotenv
requests
